// frontend/src/lib/generatedImageAssets.ts
export interface ResponsiveImageSet {
  default: string;
  srcSet: string;
  widths: number[];
}

export const shoeImageSets: Record<string, ResponsiveImageSet> = {
  'blue-nike': {
    default: blue_nike_1200,
    srcSet: `${blue_nike_320} 320w, ${blue_nike_640} 640w, ${blue_nike_1200} 1200w`,
    widths: [320, 640, 1200],
  },
  'gold-black-nike': {
    default: gold_black_nike_1200,
    srcSet: `${gold_black_nike_320} 320w, ${gold_black_nike_640} 640w, ${gold_black_nike_1200} 1200w`,
    widths: [320, 640, 1200],
  },
  'pink-nike': {
    default: pink_nike_1200,
    srcSet: `${pink_nike_320} 320w, ${pink_nike_640} 640w, ${pink_nike_1200} 1200w`,
    widths: [320, 640, 1200],
  },
  'black-brown-nike': {
    default: black_brown_nike_1200,
    srcSet: `${black_brown_nike_320} 320w, ${black_brown_nike_640} 640w, ${black_brown_nike_1200} 1200w`,
    widths: [320, 640, 1200],
  },
  'brown-pink-nike': {
    default: brown_pink_nike_1200,
    srcSet: `${brown_pink_nike_320} 320w, ${brown_pink_nike_640} 640w, ${brown_pink_nike_1200} 1200w`,
    widths: [320, 640, 1200],
  },
};

export const backgroundImageSets: Record<string, ResponsiveImageSet> = {
  'blue-bg': {
    default: blue_bg_1920,
    srcSet: `${blue_bg_1280} 1280w, ${blue_bg_1920} 1920w`,
    widths: [1280, 1920],
  },
  'gold-bg': {
    default: gold_bg_1920,
    srcSet: `${gold_bg_1280} 1280w, ${gold_bg_1920} 1920w`,
    widths: [1280, 1920],
  },
  'pink-bg': {
    default: pink_bg_1920,
    srcSet: `${pink_bg_1280} 1280w, ${pink_bg_1920} 1920w`,
    widths: [1280, 1920],
  },
  'brown-bg': {
    default: brown_bg_1920,
    srcSet: `${brown_bg_1280} 1280w, ${brown_bg_1920} 1920w`,
    widths: [1280, 1920],
  },
  'brown-pink-bg': {
    default: brown_pink_bg_1920,
    srcSet: `${brown_pink_bg_1280} 1280w, ${brown_pink_bg_1920} 1920w`,
    widths: [1280, 1920],
  },
};

export const logo_all_pages = {
    default: logo_all_pages_400,
    srcSet: `${logo_all_pages_270} 270w, ${logo_all_pages_400} 400w`,
    widths: [270, 400],
  };

export const logo_all_pages_text = {
    default: logo_all_pages_text_400,
    srcSet: `${logo_all_pages_text_200} 200w, ${logo_all_pages_text_400} 400w`,
    widths: [200, 400],
  };

export const logo_home_text = {
    default: logo_home_text_400,
    srcSet: `${logo_home_text_200} 200w, ${logo_home_text_400} 400w`,
    widths: [200, 400],
  };